+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Clean the SSH config when the terminal is disposed.
 - Keep the selected manifest in memory instead of in the workspace state.
 - Wait until the SSH server can handle connections.
+- Automatically upgrade the okteto binary if not compatible with the plugin.
 
 ## 0.1.4
 - Use `default` if there's no namespace defined in the context.